add context authprovider function add conetxt auth wrapper for app.js in index
