+++ .eslintrc.json
@@ -12,6 +12,7 @@
   "plugins": ["react"],
 
   "rules": {
+    "no-unused-vars": "off",
     "react/prop-types": "off",
     "react/jsx-key": "off"
   }
