+++ src/components/Home/Home.js
@@ -1,12 +1,16 @@
-import React from "react";
+import React, { useContext } from "react";
+import AuthContext from "../../store/auth-context";
+import Button from "../UI/Button/Button";
 
 import Card from "../UI/Card/Card";
 import classes from "./Home.module.css";
 
 const Home = () => {
+  const ctx = useContext(AuthContext);
   return (
     <Card className={classes.home}>
       <h1>Welcome back!</h1>
+      <Button onClick={ctx.onLogout}>Logout</Button>
     </Card>
   );
 };
