+++ src/store/auth-context.js
@@ -1,7 +1,39 @@
-import React from "react";
+import React, { useState, useEffect } from "react";
 
 const AuthContext = React.createContext({
   isLoggedIn: false,
+  onLogout: () => {},
+  onLogin: (email, password) => {},
 });
+
+export const AuthContextProvider = (props) => {
+  const [isLoggedIn, setIsLoggedIn] = useState(false);
+
+  useEffect(() => {
+    const storedUserPw = localStorage.getItem("isLoggedIn");
+
+    if (storedUserPw === "1") {
+      setIsLoggedIn(true);
+    }
+  }, []);
+
+  function logoutHandler() {
+    localStorage.removeItem("isLoggedIn");
+    setIsLoggedIn(false);
+  }
+
+  function loginHandler() {
+    localStorage.setItem("isLoggedIn", "1");
+    setIsLoggedIn(true);
+  }
+
+  return (
+    <AuthContext.Provider
+      value={{ isLoggedIn, onLogout: logoutHandler, onLogin: loginHandler }}
+    >
+      {props.children}
+    </AuthContext.Provider>
+  );
+};
 
 export default AuthContext;
